requirements.txt
bcrypt==3.1.3
Flask==0.12.2
Flask-Cors==3.0.3
Flask-Migrate==2.0.4
Flask-SQLAlchemy==2.2
requests==2.4.2
feedparser==5.2.1
psycopg2==2.7.3.1
python-slugify==1.2.4
feedfinder2==0.0.4
textblob==0.13.0
newspaper==0.1.0.7
socialshares==1.1.0
python-crontab==2.2.6